Rename package and minor changes
